extracted details view components
