+++ frontend/src/components/Details/ActionBar.vue
