+++ frontend/src/components/Details/AdditionalContent.vue
