+++ frontend/src/components/Details/Info/GameInfo.vue
