+++ frontend/src/components/Details/Notes.vue
