+++ frontend/src/components/Details/RelatedGames.vue
