+++ frontend/src/components/Details/States.vue
