+++ frontend/src/components/Details/Title.vue
