+++ frontend/src/components/Details/VersionSwitcher.vue
