+++ frontend/src/views/GameDetails.vue
@@ -1,16 +1,16 @@
 <script setup lang="ts">
 import EmptyGame from "@/components/common/EmptyGame.vue";
 import Cover from "@/components/common/Game/Card/Base.vue";
-import ActionBar from "@/components/common/Game/Details/ActionBar.vue";
-import AdditionalContent from "@/components/common/Game/Details/AdditionalContent.vue";
-import BackgroundHeader from "@/components/common/Game/Details/BackgroundHeader.vue";
-import FileInfo from "@/components/common/Game/Details/Info/FileInfo.vue";
-import GameInfo from "@/components/common/Game/Details/Info/GameInfo.vue";
-import Notes from "@/components/common/Game/Details/Notes.vue";
-import RelatedGames from "@/components/common/Game/Details/RelatedGames.vue";
-import Saves from "@/components/common/Game/Details/Saves.vue";
-import States from "@/components/common/Game/Details/States.vue";
-import TitleInfo from "@/components/common/Game/Details/Title.vue";
+import ActionBar from "@/components/Details/ActionBar.vue";
+import AdditionalContent from "@/components/Details/AdditionalContent.vue";
+import BackgroundHeader from "@/components/Details/BackgroundHeader.vue";
+import FileInfo from "@/components/Details/Info/FileInfo.vue";
+import GameInfo from "@/components/Details/Info/GameInfo.vue";
+import Notes from "@/components/Details/Notes.vue";
+import RelatedGames from "@/components/Details/RelatedGames.vue";
+import Saves from "@/components/Details/Saves.vue";
+import States from "@/components/Details/States.vue";
+import TitleInfo from "@/components/Details/Title.vue";
 import platformApi from "@/services/api/platform";
 import romApi from "@/services/api/rom";
 import storeDownload from "@/stores/download";
